+++ src/app/accessRequests/reviewSubServiceRequest.js
@@ -1,46 +1,14 @@
-const { getAllRequestsTypesForApprover ,getApproversForOrganisation} = require('../../infrastructure/organisations');
-const { listRolesOfService, updateUserService} = require('../../infrastructure/access');
 const {updateServiceRequest} = require('../requestService/utils');
-const { checkCacheForAllServices } = require('../../infrastructure/helpers/allServicesAppCache');
-const { getUserDetails } = require('./utils');
+const { getAllRequestsForApproval,getNewRoleDetails, getSubServiceRequestVieModel } = require('./utils');
 const { actions } = require('../constans/actions');
 const logger = require('./../../infrastructure/logger');
 const config = require('./../../infrastructure/config');
 const NotificationClient = require('login.dfe.notifications.client');
-const getAllRequestsForApproval = async (req) => {
-  const pageSize = 5;
-  const paramsSource = req.method.toUpperCase() === 'POST' ? req.body : req.query;
-  let pageNumber = parseInt(paramsSource.page, 10) || 1;
-  if (isNaN(pageNumber)) {
-    pageNumber = 1;
-  }
 
-  const allRequestsForApprover = await getAllRequestsTypesForApprover(req.user.sub, pageSize, pageNumber, req.id);
-  let { requests } = allRequestsForApprover;
-  if (requests) {
-    const userList = (await getUserDetails(requests)) || [];
 
-    requests = requests.map((user) => {
-      const userFound = userList.find((c) => c.claims.sub.toLowerCase() === user.user_id.toLowerCase());
-      const usersEmail = userFound ? userFound.claims.email : '';
-      const userName = userFound ? `${userFound.claims.given_name} ${userFound.claims.family_name}` : '';
-      return Object.assign({ usersEmail, userName }, user);
-    });
-  }
-
-  return {
-    csrfToken: req.csrfToken(),
-    title: 'Requests - DfE Sign-in',
-    currentPage: 'requests',
-    requests,
-    pageNumber,
-    numberOfPages: allRequestsForApprover.totalNumberOfPages,
-    totalNumberOfResults: allRequestsForApprover.totalNumberOfRecords,
-  };
-};
 const validate = async (req) => {
-  const buildmodel = await buildModel(req);
-  const viewModel = await extractVieModel(buildmodel, req.params.rid, req.id);
+  const buildmodel = await getEveryPendingRequest(req);
+  const viewModel = await getSubServiceRequestVieModel(buildmodel, req.params.rid, req.id);
   viewModel.selectedResponse = req.body.selectedResponse;
   const model = {
     title: 'Review request - DfE Sign-in',
@@ -60,7 +28,7 @@ const validate = async (req) => {
   }
   return model;
 };
-const buildModel = async (req) => {
+const getEveryPendingRequest = async (req) => {
   const pagedRequests = await getAllRequestsForApproval(req);
   return {
     csrfToken: req.csrfToken(),
@@ -72,39 +40,10 @@ const buildModel = async (req) => {
     totalNumberOfResults: pagedRequests.totalNumberOfResults,
   };
 };
-const getRoleAndServiceNames = async(subModel, requestId, req) => {
-let serviceId = subModel.service_id;
-let roleIds = subModel.role_ids;
-const allServices = await checkCacheForAllServices(requestId);
-const serviceDetails = allServices.services.find((x) => x.id === serviceId);
-const allRolesOfService = await getNewRoleDetails(serviceId, subModel.role_ids);
-let roleDetails = allRolesOfService.find(x => x.id === roleIds);
-if(req !== undefined){
-  req.session.role = roleDetails;
-}
-
-if(roleDetails.name)
-subModel.Role_name = roleDetails.name;
-if(serviceDetails.name)
-subModel.Service_name = serviceDetails.name;
-return subModel;
-}
-const getNewRoleDetails = async (serviceId, roleId) => {
-  return await listRolesOfService(serviceId, roleId);
-};
-const extractVieModel= async (model, rid, requestId, req) => {
-  let viewModel = model.requests.find(x => x.id === rid);
-  viewModel.csrfToken = model.csrfToken;
-  viewModel.selectedResponse= null;
-  viewModel.validationMessages= {};
-  viewModel.currentPage= 'requests';
-  viewModel = await getRoleAndServiceNames(viewModel, requestId, req);
-  return viewModel;
-};
 
 const get = async (req, res) => {
-  const model = await buildModel(req);
-  const viewModel = await extractVieModel(model, req.params.rid, req.id, req);
+  const model = await getEveryPendingRequest(req);
+  const viewModel = await getSubServiceRequestVieModel(model, req.params.rid, req.id, req);
   viewModel.backLink =  `/access-requests/requests`;
   req.session.rid = req.params.rid;
   if(req.session.roleId != undefined && req.session.roleId  !== viewModel.role_ids)
@@ -123,7 +62,6 @@ const get = async (req, res) => {
   }
   return res.render('accessRequests/views/reviewSubServiceRequest', viewModel);
 };
-
 const post = async (req, res) => {
   const model = await validate(req);
 
@@ -193,7 +131,6 @@ const post = async (req, res) => {
     return res.render('accessRequests/views/reviewSubServiceRequest', model.viewModel);
   }
 };
-
 module.exports = {
   get,
   post,
